+++ lib/Vector.h
@@ -10,7 +10,7 @@ using namespace std;
 class Vector {
     private:
         string id;
-        long double *dimension;
+        double *dimension;
         int size;
         
 
@@ -18,14 +18,14 @@ class Vector {
     public:
         Vector(string id, int size);
         ~Vector(void);
-        void initVector(long double *dimension);
+        void initVector(double *dimension);
         void printVector(void);
         int getVectorDimension(void);
-        long double getVectorItem(int index);
+        double getVectorItem(int index);
         string getVectorID(void);
         void setG(unsigned int g);
         unsigned int getG(int index);
-        long double * getVectorCoordinates();
+        double * getVectorCoordinates();
 };
 
 #endif //VECTOR
